fix: правки по ревью
