--- ewm-service/src/main/java/ru/practicum/ewm/exception/NoHeaderException.java
@@ -1,7 +1,0 @@
-package ru.practicum.ewm.exception;
-
-public class NoHeaderException extends RuntimeException {
-    public NoHeaderException(String message) {
-        super(message);
-    }
-}
